Wait for 'license-release' when calling remove().
When we call remove() on an EME session, it will raise a 'message'
event of type 'license-release'.  This should be sent to the license
server and the response given to update().  This makes DrmEngine
wait for this exchange when removing sessions.  This only applies to
'persistent-license' session types.

Note that there is a Chrome bug (https://crbug.com/690583) where calling
close() while this message is pending will cause the close() Promise
to hang. This was causing our encrypted offline test to timeout because
we didn't wait for the response.  The test is now passing.

Change-Id: I19f6e30b7bd410d33662a38484d7bc525b7db594