Override hover styles when there are touch events.
When the user touches the demo controls (e.g. on a phone), the controls
will get the :hover pseudo-class.  However, this style doesn't get
removed until the user "clicks" somewhere else.  This causes the
controls to remain visible forever.

When we are getting touch events, we should override the :hover style
and hide the controls.  However, we should only do this so long as the
clicks are from touch events.  For touchscreen laptops (that have both
touch events and a mouse), we should still allow hovering over the
controls if they used the mouse.

Closes #663

Change-Id: I2705aa8fa9e59657b6fe3603c1826125ccce950a