Adding "toString" method to HLS Tags
For HLS Drm Init Data, the whole tag needs to be passed down to
the cdm. This is a problem as by the time the drm init data is
being created the tag has been parsed and no longer has reference
to its original form.

This change adds a function "toString" to create a tag that would
be the same as its original text form.

As we create tags in our tests - it was not feasible to just give
it a text tag value so instead the function will recreate the text
form.

Change-Id: Idbc5bff3dd088a3aa3a67ef6a22749ea91bb1ccf