Don't use SPD to determine availability start.
Before we would use the suggestedPresentationDelay attribute when we
determined the availability start, which is incorrect.  This can be
seen by loading the DashIF live simulator streams and seeking to the
beginning of the seek range.

The DASH spec doesn't explicitly state this either way.  It seems to
suggest that suggestedPresentationDelay is only used to synchronize
playback and isn't used to calculate segment times or availability.
The spec also says that clients can ignore the attribute and use their
own, which suggests that it shouldn't affect availability.

Change-Id: I44d23a556c1aa40c42e2893718f8da1d3a373e8d