Call abort() before setting timestampOffset.
When appending MPEG2-TS segments, we seem to always enter a
PARSING_MEDIA_SEGMENT state in the SourceBuffer.  While we are in this
state we can't change the timestampOffset.  This means that we can't
change Periods or adapt to a different resolution.

To fix this, we call abort() before setting the timestampOffset to
reset the state.  Because abort() also resets the appendWindowEnd,
this merges the two operations so we set timestampOffset and
appendWindowEnd at the same time (since we do already).

Issue #279

Change-Id: Icd893ded9e3137179f262ba10e55e41d06ec5a77