Allow seeking to 0 in live streams
This is a slightly-temporary fix that fixes #692.  This will
be replaced by a refactoring of Playhead.  That is also why there are
no new tests for this behavior.

We calculate a "safe window" where we could seek to if we are
unbuffered.  This ensures that we don't move backward while we are
buffering and "fall" out of the availability window.  Before, we would
take the availability start and add the rebuffering goal.  This led
to a problem where we couldn't seek to 0 since we would truncate the
availability start to 0 before adding.  This meant that even if
timeShiftBufferDepth was far enough away so that 0 was "safe", we would
still not allow seeking to it.

This makes it so we add the rebuffering goal before we truncate to 0
so time 0 can be considered "safe".

This also changes how seeks before the availability window are handled.
Before, it would always seek back to the "safe" point, even if there
was enough buffered.  Now, so long as the start of the window is
buffered, we will seek to the start of the window if you seek before it.

Closes #692

Change-Id: If8fc23b4d7dd945e0e1c01dd215e998037cd4ebb