Copy playhead observing code into new PlayheadObserver.
This is the first change in a larger refactoring to split Playhead into
a new class PlayheadObserver.  This adds the new class PlayheadObserver
and the related tests.  The type isn't being used yet.  The new tests
cover similar cases as the old Playhead tests, and the old tests will
be removed once PlayheadObserver is being used.

Change-Id: I4221372da8f9b8509c824380e65a09577d8bc6ea