Allow switching between Periods with and without text
We don't allow switching between Periods that are video+audio to
video-only because it is incompatible with MSE.  When the browser
gets to the transition, it will pause waiting for the audio buffer.

However, we should allow switching between Periods that do and do not
have text streams.

Backport to v2.0.x

Issue #715

Change-Id: I77f3bf92681d8181c90169b8f4a15857b7d9e66d