Made changes to the demo page update the hash.
Now every UI element that changes a value that corresponds to a demo page
argument will also cause the hash to automatically update.
Note that this does not handle changes in the UI state based on console
input; for example, changing the debug level via console commands won't
update the selector or the hash.
Also adds an "autoplay" option to the configuration section.

Closes #709

Change-Id: I03aba29a50938848d12e0ffab32188ab96fec3b6