Don't use PTO in SegmentList with duration.
The presentationTimeOffset attribute is used to adjust the start time
of segments to the DASH timeline.  For example, if the first segment
starts at 10s, then this attribute can be used to make it start at 0.
Normally, the times in the manifest should match the times in the
segments.  So we should use PTO on the times in the manifest.

However, for SegmentList with duration we shouldn't actually use it.
According to DASH sec. 7.2.1, the MPD start time (which is approximately
the presentation time) is after PTO is applied.  In sec. 5.3.9.5.3,
when using the duration attribute, the MPD start time is calculated
from the duration (as we do).  So that time already has PTO applied.

Change-Id: I8ac3befec8a8b0cbe8dda48f5c5888e1122c5d62