Allow small gap at start when buffering.
We allow small gaps ahead of the playhead in StreamingEngine.  But this
is not mirrored in Playhead.  We should exit a buffering state even
when there is a small gap.

Note this is unrelated to the issue where Chrome won't jump gaps after
a seek.  In StreamingEngine, we don't allow this gap when seeking;
however, this change involves normal playback and doesn't involve that
bug.

Change-Id: If091994ee1a93965836e99cdb41873afb18d0fd4