Improve check for encrypted content in HLS parser.
Until we implement support for protected content for HLS, we
reject it. We used to determine it by the presence of an
'EXT-X-KEY' tag. However, it is possible for the manifest to
have the tag, but describe clear key content if the 'METHOD'
attribute of the tag is 'NONE'. This CL insures we don't reject
this kind of playlists.

Issue #279.

Change-Id: Iee853dd23116c40a43b228f0876efe373c3111af