Fix page reload when upgrading params
We used to have search parameters to control the demo app.  Now we use
hash parameters.  When upgrading, we change the URL to convert the old
parameters to the new ones.  But if the old parameters are not
present, we must be careful to avoid navigation and a page reload.

Change-Id: Ief7adb76f0dba7631752ab77aac0b28b3969f8a0