HLS: allow spaces in the codecs list.
According to the HLS spec, CODECS attribute should have a comma-
separated list of codecs. The example in the spec doesn't contain
spaces between the elements of the list.
However, I encountered spaces in the codecs list in content
encoded by Apple. It's easy to support both ways, so I decided to
do it rather than fight that battle.

Issue #279.

Change-Id: Iecb706207365a51b5964f013e32ffe13e99604c6