Remove TODO about trying to give Media Source a list of codecs.
We wanted to see if Media Source would accept a list of codecs for
a presentation without specifying which one belongs to which stream.
If it was possible, we wouldn't have to try and guess the codecs
and also wouldn't be limited to a made up list of codecs that we
know Media Source supports.
Been there, done that, doesn't work :|

Issue #279.

Change-Id: Id706022887beaa576eb0f68f2f97a278a0730f30