Handle gaps between Periods.
Now the DASH parser will remove gaps that appear between Periods.
It removes the gaps by adjusting the duration of the previous Period
to fill the gap.  Because it affects the duration, the last segment of
the Period will also be adjusted. This only affects the segment index;
if this adjustment results in gaps in the media, they will not be
jumped (yet).

Issue #555

Change-Id: Idd2cd7ad960855be01565615c8917f7191b29503