Fixed demo UI bug on Firefox.
In the current stable versions of Firefox, calling blur on a <select>
element will also close the element's associated drop-down menu.
Because we blur every element that is opened by mouse input, this made
it impossible to open up <select> element menus at all on Firefox.
To work around that, this change makes it so that mouse input no longer
blurs <select> elements.

Change-Id: I123954c8227ceb31793dbddd51376ca5eedff7a0