Fix demo page problem on Firefox.
On Firefox, the demo page wasn't loading fully; specifically, it never
ran shakaDemo.postBrowserCheckParams_. This happens if
shakaDemo.setupOfflineAssets_ returns a rejected promise.
At the moment, Firefox is failing to successfully load the offline assets,
but that's a subject for another CL.
This change makes it so that, if the offline assets fail to load, it
displays the error and continues loading the page.

Change-Id: I8f92d92ab86628f5e8555e0e10ebe51229c3afff