Added another XML parsing validity check.
Different browsers implementations of DOMParser handle invalid XML in
different ways. In some cases, instead of the DOMParser erroring it
just inserts a <parsererror> element into the resulting xml tree.
This change detects if there was a typo deeper in the manifest, such
that the outer MPD tags are intact.

Change-Id: I9b11b0762d073b8e54129cddd5f3477039f2cb39