Fix duplicate PlayheadObserver
PlayheadObserver was not being cleaned up.  This led to multiple
observers, which ultimately led to a race in which the buffering state
was broken at the end of a playback.

This bug was introduced during the Playhead/PlayheadObserver refactor,
and did not appear in an official release.  It was caught during
manual testing.

The quickest repro is to play a long video, then play a short one.
Let the short video play to the end.

The observer of the long video has a larger manifest duration value,
and therefore does not see the end of the short video as "ended".
It therefore puts us into a buffering state.

This also fixes a bug in the destroy() method itself and adds a test
to verify player teardown.

Change-Id: I60086345391e488cb1cf36aeacc5105b1566b07d