Revert "Fix quarantined StreamingEngine tests."
This reverts commit ea724fdc5cba218fb45d90e9010ee3f3c0cb5076.

The pipeline flushing code seems to cause stalls in StreamingEngine
and Player integration tests,

All gap-jumping tests triggered the debug log "Flushing media pipeline
due to stall inside buffered range", even when passing, which was a
surprise.  This would suggest that the detection of "stalls" is
incorrect.

The failing test runs would repeat this message over and over until
the test finally timed out.  This would suggest that the "flush" was
ineffective.

Reverting this seems to fix the test failures.

b/35993908

Change-Id: I476d2f97a6da7309449b1a19a54a14e21d267d89