Fix Player integration test flake.
First, by listening for the first 'timeupdate' event, we may run our
checks before the Player sees that we are playing.  Now we wait until
1 second in.  Second, there was some test contamination because of how
the timeouts were implemented.  If the tests took too long, then the
timeout for the first test would remove the listeners for the second
test, resulting in a test timeout.

b/36485220

Change-Id: Ia517cd5162cc4d8bf0aaf1b708c004c0b1145e0e