Fix StreamingEngine integration tests again.
Sometimes it is possible for the browser to get stuck inside a buffered
range.  This change will make Playhead flush the media pipeline if this
happens.  This uses a very simple heuristic to determine if we are
stalled.  It may trigger during gap jumping, but it won't have any
effect.

The difference between its previous incarnation is there is now a delay
between repeated flushing.  In Chrome and Firefox, the seek completes
fast enough to not trigger the flush again.  However in Chromecast, it
takes a few seconds to continue, meaning it would flush repeatedly
forever.  Now Playhead will only flush once every 5 seconds.

This also fixes a race in one test.  It would wait until the playhead
reached a certain time.  However, the test has the playhead start past
the time and has Playhead move it backwards.  This caused a race that
caused occasional failure or immediate completion on IE.

b/35993908

Change-Id: Id08c1f006f31530e9db74404416e6557a6242979