Fix stutter with live content on Chromecast.
When playing live streams near the beginning of the availability
window, there is a stutter.  After a seek, Chromecast takes a few
seconds to start playing, even when the content is buffered.  This
causes us to fall outside the availability window causing another seek.
This change increase the offset when seeking to help.

b/37569681

Change-Id: I6bff8a32a1009fb2c93406d46fa4bdf3264e6dfe