Restrict tracks that we don't have keys for.
There was a bug where not having a key for a track allowed it to be
played.  This was caused because we used to only pass the keys that
were updated to Player.  This meant that a missing key was
inconclusive.  But now we always pass every key we have; so if the key
isn't in the passed map, we don't have that key and we should restrict
that track.

Backported to v2.0.x

Closes #761

Change-Id: I42bf5af1e0a62d97460356bba502a859c5970e1b