Fixed a bug in how the demo detects audio assets.
In HLS, the resolution property is optional. This means we
cannot assume that a variant with a video stream has a defined
width. Such a check was causing us to incorrectly identify
variants as audio-only, and applying the audio-only poster
incorrectly.
Instead, this change checks for the presence of a video codec.

Closes #794

Change-Id: I80ebd11ebc735468220816c5133a3398436f4530