Clean up demo loader
The loader script for the demo uses document.write to add script tags.
This is not ideal, and currently generates these warnings in Chrome:

[Violation] Parser was blocked due to document.write(<script>)
[Violation] Avoid using document.write().

This replaces the import mechanism and hooks it into Closure library.

This also fixes uncompiled mode on Cobalt, which does not have
document.write.  (http://cobalt.foo/)

Fixes: 37246436

Change-Id: I4b8acf41edb314be359bca42d224112020c30709