Fix track switching when near Period boundary.
Trying to switch to a stream from a previous Period while we are
buffering a later Period can cause problems.  This is caused by
different media states "needing" different Periods.

So if we try to switch to a stream from a different Period, we should
handle the Period transition then to ensure all the streams are needing
the correct Periods.

Closes #797

Change-Id: I3d93d93bf141901fe8f8f62fd0ca9fa96a07c268