Fixes livestreams with no seek range.
When the availability window of a live stream is very narrow, in
at least some cases, the playhead can end up starting before the
availability window. In that case, the stream will fail to load unless
gap jumping is set up to jump long gaps.
This makes the playhead jump ahead if it falls behind the availability
window, to avoid that situation (and others that might have the same
effect.)

Closes #916

Change-Id: I87f8c70ba6053d3524a1546e57d55cb6528cc683