Fix filterVariants to give variants with uniform language/roles
This fixes bugs in both filterVariantsByRoleAndLanguage and
filterTextStreamsByRoleAndLanguage which would ultimately allow ABR to
adapt across streams/variants with different languages/roles. There were
multiple ways that this could happen, detailed in #947 and #918. The bug
is fixed by some re-arrangment of the filtering logic and adding logic
to the END of the filter methods that do a final pass to select a
uniform set of variant/streams.

Technically the DASH spec requires that the player not adapt in/out of a
single AdaptationSet, but since the "variant" concept does not track the
AdaptationSet that its component streams originate from, the closest
thing we can go by is the "role", since two streams with different roles
certainly come from different adaptation sets.

Fixes issue #947 and #918