Immediately switch audio/text tracks if different role selected
Previously, selecting a new audio/text track with selectAudioLanguage()
or selectTextLanguage(), with the same language but different role,
would only set the variants to switch to for ABR, but wouldn't actually
switch to any of those variants at the moment of selection due to a bug
in chooseStreams_(), which was still looking at only audio language as a
differentiating mechanism. This change honors the role as well, so that
a switch to the correct variants immediately happens.

Closes #948