Fix demo app scrubber WRT seek range
The demo scrubber used to be shaded differently for live vs VOD
content.  The introduction of configurable seek ranges for VOD
content causes the scrubber to be incorrectly shaded.

We should treat all content the same using the seek range only,
since the seek range for VOD is from 0 to duration anyway.

Closes #943

Change-Id: Ia1fd74db246e6ee0735ccf56455f0300cffc0545