Fix native control seeking on Edge
When the user seeks using Edge's native controls, Edge sets
playbackRate to 0 first.  This confused Playhead and resulted in a
playbackRate stuck at 0 after the seek.

To fix it, we filter out rate changes to 0 in the code that remembers
the previous setting during a buffering event.

Closes #951

Backported to v2.1.x

Change-Id: Ia7a9a2a6d65dcf2d74ea6fb4d92594070a1ebe6a