HlsParser: correctly parse playlist tags appearing after segments.
Media Playlist tags can appear not only before all the segments
info, but inside and after it as well. We used to assume that after
the first segment tag appears, the rest of the tags can only be
applicable to segments.
Now parser will check what kind of tag it parsed and treat it
appropriately.

Change-Id: I6b6f88c272a54f682829397c97f41bac971bce5f