Fix end of stream detection on IE & Edge
PlayheadObserver has to know when the stream has been fully buffered.
It did this based on duration and the buffered ranges, but that
required the use of a fudge factor and caused false negatives on IE &
Edge.

Instead, use MediaSource's  readyState attribute to decide when we've
buffered to the end.  This way, IE & Edge do not go into a buffering
state when the discrepancy between audio & video sizes is larger than
an arbitrary fudge factor.

Closes #913

Backported to v2.1.x

Change-Id: I307dce12a883201a88a18bf5b9a84f954a1c5033