Fix Issue #990 - Null Cue List
When a text track's mode is set to "disabled" the cues exposed cue
list is set to null. However, if you save a reference to the list,
the reference is valid. If the track is disabled, the add and remove
calls still work.

This change caches the reference to the cue's list when the track is
first created so that we can always read it - regardless of what mode
the track is set to.

Fix #990

Change-Id: I8274ea8450e664eeaa359bf9e78d13405fd2e503