Issue 991 - Subtitles Fail in Safari
It appears there are different behaviours for text cues on
safari compared to other browsers. We were using "auto" for
position, but Safari does not support setting that value for
position. The spec says that the default value is "auto", so
by not setting it, we can assume that it will be "auto".

Closes #991
Closes #1012

Change-Id: Ibedc5f01f72138df85b911bb5112d40f2f327ad3