Polyfill Safari 11 MSE bug in SourceBuffer.remove
SourceBuffer.remove, when told to remove exactly up to a keyframe,
will also remove that keyframe and all of the content till the next
keyframe. This results in far too much content being deleted, and is
breaking our tests on Safari 11.

Patching this by offseting the end of the removal range by 1ms seems
to be a viable workaround.

Closes #1048

Change-Id: I062b2dc2617f1372f885f4ccbc906d77c65412d7